fix: bug of behavior encountering broken xml
