+++ src-tauri/crates/erm/tests/write/mod.rs
@@ -1,3 +1,4 @@
+pub mod preserved_xml;
 pub mod serialize;
 mod support;
 pub mod write_snapshot;
